+++ test/music_xml_test.dart
@@ -88,11 +88,11 @@ void main() {
     group('Lyric.parse', () {
       test('with a note containing multiple lyrics', () {
         final note3 = document.score.parts.single.measures.first.notes[2];
-        expect(note3.lyrics?.first.syllabic, Syllabic.end);
+        expect(note3.lyrics?.first.syllabic, SyllabicValue.end);
         expect(note3.lyrics?.first.text, 'ny');
         expect(note3.lyrics?.first.lyricName, 'verse1');
 
-        expect(note3.lyrics?.last.syllabic, Syllabic.end);
+        expect(note3.lyrics?.last.syllabic, SyllabicValue.end);
         expect(note3.lyrics?.last.text, 're ...');
         expect(note3.lyrics?.last.lyricName, 'verse2');
       });
@@ -103,11 +103,11 @@ void main() {
         expect(lyric.lyricName, 'verse1');
 
         final firstTextItem = lyric.items.first;
-        expect(firstTextItem.syllabic, Syllabic.single);
+        expect(firstTextItem.syllabic, SyllabicValue.single);
         expect(firstTextItem.text, '1.');
 
         final secondTextItem = lyric.items.last;
-        expect(secondTextItem.syllabic, Syllabic.begin);
+        expect(secondTextItem.syllabic, SyllabicValue.begin);
         expect(secondTextItem.text, 'Ma');
       });